python_service/requirements.txt
flask
flask-cors
pillow
python-dotenv

transformers
torch
